Created an enemy to test with.
Made an enemy class.  Made an enemy swordsman class as a test and
rendered one enemy to mess around with.  I am going to see if I can add
the attack animation and basic attacking sequencing when one of the
heros and the enemies come into contact with each other.