Added different Enemies!
Added an axeman, assasin(ninja), musketeer, and spearman.  We don't have
to use all of these enemy classes for the sake of space, but I think the
game will be more fun with enemies with a variety of powers and
abilities.  Also, it makes sense that they are human soldiers as opposed
to undead skeletons.